refactor gradscaler for amp
