+++ torch/amp/__init__.py
@@ -1,1 +1,2 @@
 from .autocast_mode import autocast
+from .grad_scaler import GradScaler
